requirements.txt
flask
flask-cors
pandas
scikit-learn
joblib
google-genai
python-dotenv
gunicorn